requirements/pip.txt
boto3==1.28.63
botocore==1.31.63
celery==5.3.1
channels==4.2.0
daphne==4.0.0
dj-database-url==1.0.0
Django==5.1.3
django-cors-headers==4.3.0
django-crontab==0.7.1
django-decouple==2.1
django-filter==23.3
django-rest-passwordreset==1.5.0
django-storages==1.14.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
google-api-core==2.15.0
google-api-python-client==2.93.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
numpy==1.26.1
pandas==2.1.2
Pillow==11.0.0
psycopg2-binary==2.9.9
redis==5.2.0
requests==2.31.0
xhtml2pdf==0.2.10
xmltodict==0.14.0
openpyxl==3.1.5
